# web/tooling/openapi/workforce.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Workforce.Host.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5041/"
    }
  ],
  "paths": {
...
    "/api/v1/settlement/batches/generate": {
      "post": {
        "tags": [
          "Settlement"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateSettlementBatchRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateSettlementBatchRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateSettlementBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementBatchDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementBatchDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementBatchDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/settlement/batches/{id}": {
      "get": {
        "tags": [
          "Settlement"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementBatchDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementBatchDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementBatchDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settlement/batches/{id}/approve": {
      "post": {
        "tags": [
          "Settlement"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveBatchRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveBatchRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/settlement/batches/{id}/export": {
      "post": {
        "tags": [
          "Settlement"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApproveBatchRequest": {
        "required": [
          "expectedRowVersion"
        ],
        "type": "object",
        "properties": {
          "expectedRowVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "uint32"
          }
        }
      },
...
      "GenerateSettlementBatchRequest": {
        "required": [
          "payrollRunId",
          "batchNumber",
          "paymentDate"
        ],
        "type": "object",
        "properties": {
          "payrollRunId": {
            "type": "string",
            "format": "uuid"
          },
          "batchNumber": {
            "type": "string"
          },
          "paymentDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "PaymentInstructionDto": {
        "required": [
          "id",
          "employmentId",
          "beneficiaryName",
          "bankCode",
          "accountMasked",
          "amount",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "employmentId": {
            "type": "string",
            "format": "uuid"
          },
          "beneficiaryName": {
            "type": "string"
          },
          "bankCode": {
            "type": "string"
          },
          "accountMasked": {
            "type": "string"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "SettlementBatchDetailDto": {
        "required": [
          "id",
          "payrollRunId",
          "batchNumber",
          "totalAmount",
          "currency",
          "paymentDate",
          "status",
          "rowVersion",
          "instructions"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "payrollRunId": {
            "type": "string",
            "format": "uuid"
          },
          "batchNumber": {
            "type": "string"
          },
          "totalAmount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "currency": {
            "type": "string"
          },
          "paymentDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string"
          },
          "rowVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "uint32"
          },
          "instructions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentInstructionDto"
            }
          }
        }
      },
      "SettlementBatchDto": {
        "required": [
          "id",
          "payrollRunId",
          "batchNumber",
          "totalAmount",
          "currency",
          "paymentDate",
          "status",
          "instructionCount",
          "rowVersion"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "payrollRunId": {
            "type": "string",
            "format": "uuid"
          },
          "batchNumber": {
            "type": "string"
          },
          "totalAmount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "currency": {
            "type": "string"
          },
          "paymentDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string"
          },
          "instructionCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "rowVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "uint32"
          }
        }
      },
...
    }
  },
...
}